refactor : rename file and add methods to map DTO to Domain
